% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_bioconductor_packages}
\alias{available_bioconductor_packages}
\title{Available bioconductor packages}
\format{
a list of packages
}
\usage{
available_bioconductor_packages
}
\description{
Contains a list of all the packages that are available in bioconductor
}
\keyword{datasets}
