% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requirements.R
\name{create_requirements}
\alias{create_requirements}
\title{Create requirements file}
\usage{
create_requirements(project_dir, output_dir)
}
\arguments{
\item{project_dir}{path to project directory}

\item{output_dir}{path to output directory where requirements file will be saved}
}
\value{
dataframe with requirements
}
\description{
Create requirements file
}
\examples{
examples create_requirements("~/Coding/R", "~/Coding/reqs")
}
