% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{extract_pkg_from_script}
\alias{extract_pkg_from_script}
\title{Extract package from script}
\usage{
extract_pkg_from_script(pkg_loader, txt)
}
\arguments{
\item{pkg_loader}{package loader (e.g. library, pacman, require)}

\item{txt}{text to extract package from}
}
\value{
packages
}
\description{
Extract package from script
}
