% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{find_pkg_for_function}
\alias{find_pkg_for_function}
\title{Find package of the function}
\usage{
find_pkg_for_function(function_name, lib.loc = .libPaths())
}
\arguments{
\item{function_name}{Name of the function.}

\item{lib.loc}{Location of the packages that are installed.}
}
\value{
List of the packages where the given function exists
}
\description{
Uses the name of a function to look for the package.
}
