% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{get_explicit_pkgs}
\alias{get_explicit_pkgs}
\title{Get explicit packages that are used}
\usage{
get_explicit_pkgs(path_to_script)
}
\arguments{
\item{path_to_script}{path to script for which you want the functions}
}
\value{
list of packages
}
\description{
Get explicit packages that are used
}
