% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_functions_from_unknown_pkgs}
\alias{get_functions_from_unknown_pkgs}
\title{Get functions from unknown packages}
\usage{
get_functions_from_unknown_pkgs(path_to_script)
}
\arguments{
\item{path_to_script}{path to script for which you want the functions}
}
\value{
list of functions that are not from an existing package
}
\description{
Get functions from unknown packages
}
