% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_requirements.R
\name{get_github_repo}
\alias{get_github_repo}
\title{Get github repository from url}
\usage{
get_github_repo(gh_url)
}
\arguments{
\item{gh_url}{url to github repository}
}
\value{
github repository
}
\description{
Get github repository from url
}
