% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages.R
\name{get_implicit_pkgs}
\alias{get_implicit_pkgs}
\title{Get list of implicit packages}
\source{
https://stackoverflow.com/questions/59725527/list-all-the-packages-required-in-a-script-assuming-packagefunction-in-r
}
\usage{
get_implicit_pkgs(path_to_script)
}
\arguments{
\item{path_to_script}{path to Rscript for which you want to obtain the
implicitly called packges}
}
\value{
return
}
\description{
Get list of implicit packages
}
\examples{
examples
}
