% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{get_installed_pkgs}
\alias{get_installed_pkgs}
\title{Get installed packages}
\usage{
get_installed_pkgs(libpath = .libPaths())
}
\arguments{
\item{libpath}{path to library}
}
\value{
data frame with installed packages
}
\description{
Get installed packages
}
\examples{
get_installed_pkgs()
}
