% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{get_pkgs}
\alias{get_pkgs}
\title{Get packages}
\usage{
get_pkgs(pkgs = c("loaded", "attached", "installed")[1], include_base = TRUE)
}
\arguments{
\item{pkgs}{category of packages to obtain: loaded (default), attached or installed}
}
\value{
list of packages
}
\description{
Get packages
}
\examples{
get_pkgs("loaded")
}
