% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_user_functions}
\alias{get_user_functions}
\title{Get functions defined by user}
\usage{
get_user_functions(
  path_to_script = NULL,
  pkgs_to_incl = c("loaded", "attached", "installed")[3]
)
}
\arguments{
\item{path_to_script}{path to script for which you want the functions}

\item{pkgs_to_incl}{packages to include}
}
\value{
list of functions that are not from an existing package
}
\description{
Given a script, extracts all the functions and checks whether these functions
are defined by the user. This is done by looking up whether the function is
part of package that is either loaded, attached or installed.
}
\examples{
examples
}
