% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_github}
\alias{is_github}
\title{Check if package is from GitHub}
\usage{
is_github(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
TRUE if package is from GitHub
}
\description{
Check if package is from GitHub
}
\examples{
is_github("tidyverse") # FALSE
is_github("tidyverse/tidyverse") # TRUE
}
