% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_github_pkg}
\alias{is_github_pkg}
\title{Check if package is from GitHub}
\usage{
is_github_pkg(pkg, libpath)
}
\arguments{
\item{pkg}{package name}

\item{libpath}{path to library}
}
\value{
TRUE if package is from GitHub
}
\description{
Check if package is from GitHub
}
\examples{
is_github_pkg("tidyverse") # FALSE
is_github("tidyverse/tidyverse") # TRUE
}
