% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_matched_pkg}
\alias{is_matched_pkg}
\title{Valid package}
\usage{
is_matched_pkg(pkg, installed_pkgs)
}
\arguments{
\item{pkg}{package name}

\item{installed_pkgs}{dataframe from get_installed_pkgs()}
}
\value{
package name without version or error if package is not valid
}
\description{
Check whether a package is valid based on dataframe from get_installed_pkgs.
}
\examples{
is_matched_pkg("tidyverse", c("tidyverse", "dplyr")) # "tidyverse"
}
