% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkgs_cran}
\alias{pkgs_cran}
\title{Available CRAN packages}
\format{
a list of packages
}
\usage{
pkgs_cran
}
\description{
Contains a list of all the packages that are available in CRAN
}
\keyword{datasets}
