% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reqs.R
\name{create_reqs}
\alias{create_reqs}
\title{Create requirements file}
\usage{
create_reqs(
  project_dir,
  output_dir = NULL,
  libpath = .libPaths(),
  return_path = TRUE,
  is_offline = TRUE
)
}
\arguments{
\item{project_dir}{path to project directory}

\item{output_dir}{path to output directory where requirements file will be saved}
}
\value{
dataframe with requirements
}
\description{
Create requirements file
}
\examples{
examples create_reqs("~/Coding/R", "~/Coding/reqs")
}
