% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gh_url}
\alias{get_gh_url}
\title{Get GitHub URL for package}
\usage{
get_gh_url(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
GitHub URL for package
}
\description{
Get GitHub URL for package
}
\examples{
get_gh_url("tidyverse") # NULL
}
