% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_reqs.R
\name{install_from_reqs}
\alias{install_from_reqs}
\title{Install packages from a requirements file}
\usage{
install_from_reqs(req_path, libpath)
}
\arguments{
\item{req_path}{path to requirements file}
}
\description{
Install packages from a requirements file
}
